+++ cpp/__cplusplus.cpp
@@ -1,0 +1,1 @@
+../macro/__cplusplus.cpp