+++ cpp/format.cpp
@@ -1,0 +1,13 @@
+#include <format>
+#include <iostream>
+
+std::string format(void)
+{
+	return std::format("$${}_{}", "Hello", "World");
+}
+
+int main(void)
+{
+	std::cout << format() << std::endl;
+	return 0;
+}